+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name">SenacPlanner</string>
+    <string name="app_name">TelaLogin</string>
+
+    <string name="login">Login</string>
+    <string name="senha">Senha</string>
+    <string name="entrar">Entrar</string>
 </resources>
