+++ app/src/main/res/values/themes.xml
@@ -1,9 +1,10 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.SenacPlanner" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.TelaLogin" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
+        <item name="colorControlActivated">#004A8D</item>
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.SenacPlanner" parent="Base.Theme.SenacPlanner" />
+    <style name="Theme.TelaLogin" parent="Base.Theme.TelaLogin" />
 </resources>
